+++ packages/datadog_flutter_plugin/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Cancel spans on DatadogTrackingHttpClient when RUM is enabled (prevent spans
+  from leaking native resources)
+
 ## 1.0.0-alpha.1
 
 * Support for Logging, Tracing (including Datadog Distributed Tracing) and RUM